Add logic contract tests, update doc files

+++ contracts/mocks/Slice.sol
@@ -13,7 +13,7 @@ contract Slice is OwnableUpgradeable, ERC20Upgradeable, ISlice {
     function initialize(uint256 _initialSupply) public initializer {
         OwnableUpgradeable.__Ownable_init();
         ERC20Upgradeable.__ERC20_init_unchained("SLICE", "SLC");
-        _mint(msg.sender, _initialSupply.mul(uint(1e18)));
+        _mint(msg.sender, _initialSupply.mul(1e18));
     }
 
     /**
